pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "edwa1401", email = "edwa14011977@gmail.com"},
]
dependencies = [
    "requests>=2.31.0",
    "django>=4.2.6",
    "django-extensions>=3.2.3",
    "django-stubs>=4.2.5",
    "django-filter>=23.3",
    "openpyxl>=3.1.2",
    "beautifulsoup4>=4.12.2",
    "xlrd>=2.0.1",
    "types-requests>=2.31.0.10",
    "types-beautifulsoup4>=4.12.0.7",
    "orjson>=3.9.10",
    "faker>=20.0.0",
    "celery[redis]>=5.3.6",
    "pydantic>=2.5.2",
    "redis>=5.0.1",
    "django-celery-results>=2.5.1",
    "django-celery-beat>=2.5.0",
    "django-redis>=5.4.0",
    "backoff>=2.2.1",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.1.16",
    "gunicorn>=21.2.0",
    "djangorestframework>=3.14.0",
    "markdown>=3.5.2",
    "djangorestframework-stubs>=3.14.5",
    "django-url-filter>=0.3.15",
    "django-debug-toolbar>=4.2.0",
    "django-typer>=0.4.0b0",
    "django-crispy-forms>=2.1",
    "crispy-bootstrap4>=2023.1",
    "drf-yasg>=1.21.7",
    "mkdocs>=1.5.3",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.mypy]
plugins = ['mypy_django_plugin.main', 'mypy_drf_plugin.main']
ignore_missing_imports = false
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = '''(?x) (env\.|^migrations\.|^manage\.py$|^conftest\.py$|^test\.|^tests\.)'''


[[tool.mypy.overrides]]
module = 'petroleum_prices.settings'
disable_error_code = 'var-annotated'


[[tool.mypy.overrides]]
module = "django_extensions.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "django_filters.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "drf_yasg.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "xlrd.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "django_typer.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "conftest.py"
disallow_untyped_defs = false


[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false


[tool.django-stubs]
django_settings_module = 'petroleum_prices.settings'


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py", "test*.py"]
addopts = [
    '--force-testdox',
    '--cov=["prices_analyzer", "rail_tariff", "spimex_parser", "rest_api"]',
    '--cov-report=xml:coverage.xml',
    '--cov-branch',
    '--cov-fail-under=60',
    ]


[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
lint.select = ["E4", "E7", "E9", "F", "B", "ERA001"]
lint.ignore = ["ERA001"]

# Allow fix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
    "settings.py",
    "test_*.py",
    "conftest.py",
    ]


# Same as Black.
line-length = 100


# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


# Assume Python 3.11
target-version = 'py311'


[tool.ruff.format]
# 5. Use single quotes for non-triple-quoted strings.
quote-style = "single"


[tool.pdm.dev-dependencies]
test = [
    "pytest-testdox>=3.1.0",
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
]
linter = [
    "ruff>=0.1.5",
]
dev = [
    "ipython>=8.17.2",
    "tree>=0.2.4",
    "mypy>=1.7.0",
    "typer[all]>=0.9.0",
]
